requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.6.1
monotonic==1.6
multidict==6.1.0
numpy==2.2.2
openbb-core==1.4.0
pandas==2.2.3
posthog==3.12.0
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.18
pytz==2025.1
requests==2.32.3
ruff==0.7.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uuid7==0.1.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==14.2
yarl==1.18.3
zipp==3.21.0
